Adapt to deploy on Render
