+++ server/package.json
@@ -5,9 +5,8 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
-    "start": "node api/index.js",
-    "dev": "nodemon api/index.js",
-    "vercel-build": "echo hello"
+    "start": "index.js",
+    "dev": "nodemon index.js"
   },
   "author": "",
   "license": "ISC",
